#  Remove reflection usage for SKObject creation