--- a/binding/Binding/GRContext.cs
+++ b/binding/Binding/GRContext.cs
@@ -40,7 +40,7 @@ namespace SkiaSharp
 			backend switch
 			{
 				GRBackend.Metal => throw new NotSupportedException (),
-				GRBackend.OpenGL => GetObject<GRContext> (SkiaApi.gr_context_make_gl (backendContext)),
+				GRBackend.OpenGL => GetObject (SkiaApi.gr_context_make_gl (backendContext)),
 				GRBackend.Vulkan => throw new NotSupportedException (),
 				_ => throw new ArgumentOutOfRangeException (nameof (backend)),
 			};
@@ -51,7 +51,7 @@ namespace SkiaSharp
 			CreateGl (null);
 
 		public static GRContext CreateGl (GRGlInterface backendContext) =>
-			GetObject<GRContext> (SkiaApi.gr_context_make_gl (backendContext == null ? IntPtr.Zero : backendContext.Handle));
+			GetObject (SkiaApi.gr_context_make_gl (backendContext == null ? IntPtr.Zero : backendContext.Handle));
 
 		//
 
@@ -104,5 +104,7 @@ namespace SkiaSharp
 		[EditorBrowsable (EditorBrowsableState.Never)]
 		[Obsolete]
 		public int GetRecommendedSampleCount (GRPixelConfig config, float dpi) => 0;
+
+		internal static GRContext GetObject (IntPtr handle) => TryGetObject<GRContext> (handle, out var obj) ? obj : new GRContext (handle, true);
 	}
 }