--- a/binding/Binding/GRContext.cs
+++ b/binding/Binding/GRContext.cs
@@ -105,6 +105,6 @@ namespace SkiaSharp
 		[Obsolete]
 		public int GetRecommendedSampleCount (GRPixelConfig config, float dpi) => 0;
 
-		internal static GRContext GetObject (IntPtr handle) => TryGetObject<GRContext> (handle, out var obj) ? obj : new GRContext (handle, true);
+		internal static GRContext GetObject (IntPtr handle) => GetOrAddObject (handle, (h,o) => new GRContext (h, o));
 	}
 }