--- a/binding/Binding/HandleDictionary.cs
+++ b/binding/Binding/HandleDictionary.cs
@@ -1,21 +1,16 @@
 ﻿using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Reflection;
 using System.Threading;
 
 namespace SkiaSharp
 {
 	internal static class HandleDictionary
 	{
-		private static readonly Type IntPtrType = typeof (IntPtr);
-		private static readonly Type BoolType = typeof (bool);
 
 #if THROW_OBJECT_EXCEPTIONS
 		internal static readonly ConcurrentBag<Exception> exceptions = new ConcurrentBag<Exception> ();
 #endif
-
-		internal static readonly ConcurrentDictionary<Type, ConstructorInfo> constructors = new ConcurrentDictionary<Type, ConstructorInfo> ();
 		internal static readonly Dictionary<IntPtr, WeakReference> instances = new Dictionary<IntPtr, WeakReference> ();
 
 		internal static readonly ReaderWriterLockSlim instancesLock = new ReaderWriterLockSlim ();
@@ -40,59 +35,37 @@ namespace SkiaSharp
 			}
 		}
 
-		/// <summary>
-		/// Retrieve or create an instance for the native handle.
-		/// </summary>
-		/// <returns>The instance, or null if the handle was null.</returns>
-		internal static TSkiaObject GetObject<TSkiaObject, TSkiaImplementation> (IntPtr handle, bool owns = true, bool unrefExisting = true, bool refNew = false)
+		internal static bool TryGetObject<TSkiaObject> (IntPtr handle, bool owns, bool unrefExisting, bool refNew, out TSkiaObject obj)
 			where TSkiaObject : SKObject
-			where TSkiaImplementation : SKObject, TSkiaObject
 		{
-			if (handle == IntPtr.Zero)
-				return null;
+			if (handle == IntPtr.Zero) {
+				obj = null;
+				return true;
+			}
 
 			instancesLock.EnterUpgradeableReadLock ();
+
 			try {
-				if (GetInstanceNoLocks<TSkiaObject> (handle, out var instance)) {
+				if (GetInstanceNoLocks<TSkiaObject> (handle, out obj)) {
 					// some object get automatically referenced on the native side,
 					// but managed code just has the same reference
-					if (unrefExisting && instance is ISKReferenceCounted refcnt) {
+					if (unrefExisting && obj is ISKReferenceCounted refcnt) {
 #if THROW_OBJECT_EXCEPTIONS
 						if (refcnt.GetReferenceCount () == 1)
 							throw new InvalidOperationException (
 								$"About to unreference an object that has no references. " +
-								$"H: {handle.ToString ("x")} Type: {instance.GetType ()}");
+								$"H: {handle.ToString ("x")} Type: {obj.GetType ()}");
 #endif
 						refcnt.SafeUnRef ();
 					}
 
-					return instance;
+					return true;
 				}
-
-				var type = typeof (TSkiaImplementation);
-				var constructor = constructors.GetOrAdd (type, t => GetConstructor (t));
-
-				// we don't need to go into a writable here as the object will do it in the Handle property
-				var obj = (TSkiaObject)constructor.Invoke (new object[] { handle, owns });
-				if (refNew && obj is ISKReferenceCounted toRef)
-					toRef.SafeRef ();
-				return obj;
 			} finally {
 				instancesLock.ExitUpgradeableReadLock ();
 			}
 
-			static ConstructorInfo GetConstructor (Type type)
-			{
-				var ctors = type.GetTypeInfo ().DeclaredConstructors;
-
-				foreach (var ctor in ctors) {
-					var param = ctor.GetParameters ();
-					if (param.Length == 2 && param[0].ParameterType == IntPtrType && param[1].ParameterType == BoolType)
-						return ctor;
-				}
-
-				throw new MissingMethodException ($"No constructor found for {type.FullName}.ctor(System.IntPtr, System.Boolean)");
-			}
+			return false;
 		}
 
 		/// <summary>