--- a/binding/Binding/SKCodec.cs
+++ b/binding/Binding/SKCodec.cs
@@ -312,7 +312,7 @@ namespace SkiaSharp
 				throw new ArgumentNullException (nameof (stream));
 
 			fixed (SKCodecResult* r = &result) {
-				var codec = GetObject<SKCodec> (SkiaApi.sk_codec_new_from_stream (stream.Handle, r));
+				var codec = GetObject (SkiaApi.sk_codec_new_from_stream (stream.Handle, r));
 				stream.RevokeOwnership (codec);
 				return codec;
 			}
@@ -325,7 +325,7 @@ namespace SkiaSharp
 			if (data == null)
 				throw new ArgumentNullException (nameof (data));
 
-			return GetObject<SKCodec> (SkiaApi.sk_codec_new_from_data (data.Handle));
+			return GetObject (SkiaApi.sk_codec_new_from_data (data.Handle));
 		}
 
 		// utils
@@ -343,5 +343,7 @@ namespace SkiaSharp
 				return new SKFrontBufferedManagedStream (stream, MinBufferedBytesNeeded, true);
 			}
 		}
+
+		internal static SKCodec GetObject (IntPtr handle) => TryGetObject<SKCodec> (handle, out var obj) ? obj : new SKCodec (handle, true);
 	}
 }