--- a/binding/Binding/SKColorFilter.cs
+++ b/binding/Binding/SKColorFilter.cs
@@ -20,12 +20,12 @@ namespace SkiaSharp
 
 		public static SKColorFilter CreateBlendMode(SKColor c, SKBlendMode mode)
 		{
-			return GetObject<SKColorFilter>(SkiaApi.sk_colorfilter_new_mode((uint)c, mode));
+			return GetObject (SkiaApi.sk_colorfilter_new_mode((uint)c, mode));
 		}
 
 		public static SKColorFilter CreateLighting(SKColor mul, SKColor add)
 		{
-			return GetObject<SKColorFilter>(SkiaApi.sk_colorfilter_new_lighting((uint)mul, (uint)add));
+			return GetObject (SkiaApi.sk_colorfilter_new_lighting((uint)mul, (uint)add));
 		}
 
 		public static SKColorFilter CreateCompose(SKColorFilter outer, SKColorFilter inner)
@@ -34,7 +34,7 @@ namespace SkiaSharp
 				throw new ArgumentNullException(nameof(outer));
 			if (inner == null)
 				throw new ArgumentNullException(nameof(inner));
-			return GetObject<SKColorFilter>(SkiaApi.sk_colorfilter_new_compose(outer.Handle, inner.Handle));
+			return GetObject (SkiaApi.sk_colorfilter_new_compose(outer.Handle, inner.Handle));
 		}
 
 		public static SKColorFilter CreateColorMatrix(float[] matrix)
@@ -44,13 +44,13 @@ namespace SkiaSharp
 			if (matrix.Length != 20)
 				throw new ArgumentException("Matrix must have a length of 20.", nameof(matrix));
 			fixed (float* m = matrix) {
-				return GetObject<SKColorFilter> (SkiaApi.sk_colorfilter_new_color_matrix (m));
+				return GetObject (SkiaApi.sk_colorfilter_new_color_matrix (m));
 			}
 		}
 
 		public static SKColorFilter CreateLumaColor()
 		{
-			return GetObject<SKColorFilter>(SkiaApi.sk_colorfilter_new_luma_color());
+			return GetObject (SkiaApi.sk_colorfilter_new_luma_color());
 		}
 
 		public static SKColorFilter CreateTable(byte[] table)
@@ -60,7 +60,7 @@ namespace SkiaSharp
 			if (table.Length != TableMaxLength)
 				throw new ArgumentException($"Table must have a length of {TableMaxLength}.", nameof(table));
 			fixed (byte* t = table) {
-				return GetObject<SKColorFilter> (SkiaApi.sk_colorfilter_new_table (t));
+				return GetObject (SkiaApi.sk_colorfilter_new_table (t));
 			}
 		}
 
@@ -79,18 +79,20 @@ namespace SkiaSharp
 			fixed (byte* r = tableR)
 			fixed (byte* g = tableG)
 			fixed (byte* b = tableB) {
-				return GetObject<SKColorFilter> (SkiaApi.sk_colorfilter_new_table_argb (a, r, g, b));
+				return GetObject (SkiaApi.sk_colorfilter_new_table_argb (a, r, g, b));
 			}
 		}
 
 		public static SKColorFilter CreateHighContrast(SKHighContrastConfig config)
 		{
-			return GetObject<SKColorFilter>(SkiaApi.sk_colorfilter_new_high_contrast(&config));
+			return GetObject (SkiaApi.sk_colorfilter_new_high_contrast(&config));
 		}
 
 		public static SKColorFilter CreateHighContrast(bool grayscale, SKHighContrastConfigInvertStyle invertStyle, float contrast)
 		{
 			return CreateHighContrast(new SKHighContrastConfig(grayscale, invertStyle, contrast));
 		}
+
+		internal static SKColorFilter GetObject (IntPtr handle) => TryGetObject<SKColorFilter> (handle, out var obj) ? obj : new SKColorFilter (handle, true);
 	}
 }