--- a/binding/Binding/SKData.cs
+++ b/binding/Binding/SKData.cs
@@ -48,7 +48,7 @@ namespace SkiaSharp
 		{
 			if (SizeOf <IntPtr> () == 4 && length > UInt32.MaxValue)
 				throw new ArgumentOutOfRangeException (nameof (length), "The length exceeds the size of pointers.");
-			return GetObject<SKData> (SkiaApi.sk_data_new_with_copy ((void*)bytes, (IntPtr) length));
+			return GetObject (SkiaApi.sk_data_new_with_copy ((void*)bytes, (IntPtr) length));
 		}
 
 		public static SKData CreateCopy (byte[] bytes) =>
@@ -57,7 +57,7 @@ namespace SkiaSharp
 		public static SKData CreateCopy (byte[] bytes, ulong length)
 		{
 			fixed (byte* b = bytes) {
-				return GetObject<SKData> (SkiaApi.sk_data_new_with_copy (b, (IntPtr)length));
+				return GetObject (SkiaApi.sk_data_new_with_copy (b, (IntPtr)length));
 			}
 		}
 
@@ -72,7 +72,7 @@ namespace SkiaSharp
 
 		public static SKData Create (int size)
 		{
-			return GetObject<SKData> (SkiaApi.sk_data_new_uninitialized ((IntPtr) size));
+			return GetObject (SkiaApi.sk_data_new_uninitialized ((IntPtr) size));
 		}
 
 		public static SKData Create (ulong size)
@@ -80,7 +80,7 @@ namespace SkiaSharp
 			if (SizeOf <IntPtr> () == 4 && size > UInt32.MaxValue)
 				throw new ArgumentOutOfRangeException (nameof (size), "The size exceeds the size of pointers.");
 				
-			return GetObject<SKData> (SkiaApi.sk_data_new_uninitialized ((IntPtr) size));
+			return GetObject (SkiaApi.sk_data_new_uninitialized ((IntPtr) size));
 		}
 
 		public static SKData Create (string filename)
@@ -90,7 +90,7 @@ namespace SkiaSharp
 
 			var utf8path = StringUtilities.GetEncodedText (filename, SKTextEncoding.Utf8);
 			fixed (byte* u = utf8path) {
-				return GetObject<SKData> (SkiaApi.sk_data_new_from_file (u));
+				return GetObject (SkiaApi.sk_data_new_from_file (u));
 			}
 		}
 
@@ -141,7 +141,7 @@ namespace SkiaSharp
 			if (stream == null)
 				throw new ArgumentNullException (nameof (stream));
 
-			return GetObject<SKData> (SkiaApi.sk_data_new_from_stream (stream.Handle, (IntPtr) length));
+			return GetObject (SkiaApi.sk_data_new_from_stream (stream.Handle, (IntPtr) length));
 		}
 
 		public static SKData Create (SKStream stream, ulong length)
@@ -149,7 +149,7 @@ namespace SkiaSharp
 			if (stream == null)
 				throw new ArgumentNullException (nameof (stream));
 
-			return GetObject<SKData> (SkiaApi.sk_data_new_from_stream (stream.Handle, (IntPtr) length));
+			return GetObject (SkiaApi.sk_data_new_from_stream (stream.Handle, (IntPtr) length));
 		}
 
 		public static SKData Create (SKStream stream, long length)
@@ -157,7 +157,7 @@ namespace SkiaSharp
 			if (stream == null)
 				throw new ArgumentNullException (nameof (stream));
 
-			return GetObject<SKData> (SkiaApi.sk_data_new_from_stream (stream.Handle, (IntPtr) length));
+			return GetObject (SkiaApi.sk_data_new_from_stream (stream.Handle, (IntPtr) length));
 		}
 
 		public static SKData Create (IntPtr address, int length)
@@ -176,7 +176,7 @@ namespace SkiaSharp
 				? new SKDataReleaseDelegate ((addr, _) => releaseProc (addr, context))
 				: releaseProc;
 			var proxy = DelegateProxies.Create (del, DelegateProxies.SKDataReleaseDelegateProxy, out _, out var ctx);
-			return GetObject<SKData> (SkiaApi.sk_data_new_with_proc ((void*)address, (IntPtr)length, proxy, (void*)ctx));
+			return GetObject (SkiaApi.sk_data_new_with_proc ((void*)address, (IntPtr)length, proxy, (void*)ctx));
 		}
 
 		internal static SKData FromCString (string str)
@@ -195,7 +195,7 @@ namespace SkiaSharp
 				if (offset > UInt32.MaxValue)
 					throw new ArgumentOutOfRangeException (nameof (offset), "The offset exceeds the size of pointers.");
 			}
-			return GetObject<SKData> (SkiaApi.sk_data_new_subset (Handle, (IntPtr) offset, (IntPtr) length));
+			return GetObject (SkiaApi.sk_data_new_subset (Handle, (IntPtr) offset, (IntPtr) length));
 		}
 
 		// ToArray
@@ -253,6 +253,8 @@ namespace SkiaSharp
 			}
 			GC.KeepAlive (this);
 		}
+
+		internal static SKData GetObject (IntPtr handle) => TryGetObject<SKData> (handle, out var obj) ? obj : new SKData (handle, true);
 
 		//