**Description of Change**

Describe your changes here. 

**Bugs Fixed**

- Related to issue #

Provide links to issues here. Ensure that a GitHub issue was created for your feature or bug fix before sending PR.

**API Changes**

List all API changes here (or just put None), example:

Added: 
 
- `string Class.Property { get; set; }`
- `void Class.Method();`

Changed:

 - `object Cell.OldPropertyName => object Cell.NewPropertyName`

**Behavioral Changes**

Describe any non-bug related behavioral changes that may change how users app behaves when upgrading to this version of the codebase.

**PR Checklist**

- [ ] Has tests (if omitted, state reason in description)
- [x] Rebased on top of master at time of PR
- [ ] Changes adhere to coding standard
- [ ] Updated documentation