@mattleibow 2020-04-03
Just a unit test failure due... because we no longer have to have a test for a field:

         D:\a\1\s\tests\Tests\GarbageCleanupFixture.cs(23,34): error CS0117: 'HandleDictionary' does not contain a definition for 'constructors' [D:\a\1\s\tests\SkiaSharp.Desktop.Tests\SkiaSharp.Desktop.Tests.csproj]