@mattleibow 2020-04-03
Pity there is no generic constraint for complex constructors `new(IntPtr, bool)` 😢