@Gillibald 2020-04-03
Yep but we could just introduce parameterless ctor and move the current logic to a internal virtual method I guess.