@mattleibow 2020-04-03
Hmmmm. We could do that.

Already the code is set up to support setting the handle later in the construction. So technically, we can do:

```csharp
var obj = new SKSomething();
obj.SetHandle(handle: handle, owns: true);
```

The issue is with already-default constructors, like `new SKPaint()`. That already constructs a new native object. From user code.

Not sure how to deal with that...