@Gillibald 2020-04-03
Might a special paint help here? `ReferencedSKPaint : SKPaint` that is used by code that wants to expose a already existing instance?