@Gillibald 2020-04-05
Okay. A parameterless ctor will not work because it has to be public. So the delegate approach is the best we can get for now without duplicating code.