@mattleibow 2020-04-05
What project were you using to make these changes? Was it the source project? I managed to fix the unit tests in VS2019 (I think) so it might also be cool to just open the .NET Core tests and then use them as you work on things. And you can use it to test chunks of code. Not sure what is happening with the full framework test not running in VS, but it work on the command line so no biggie.