@Gillibald 2020-04-05
I am working with the Netcore Tests project. That isn't including all tests.