@Gillibald 2020-04-05
Do we get a public ci build again that let's us see the failing tests?