@mattleibow 2020-04-07
We were working on that, but we also keeping our usage of DevOps down so we can let the COVID-19 responders get priority. 

_I say we smack this virus a big one and scare nature into a state where it doesn't want to pop another for a very long time 😆 (my expert medical opinion)_