@mattleibow 2020-04-09
DevOps. Sometimes it just fails to download the Windows SDK or some random thing. Restarting it. Sometimes the asset copy between stages also just forgets to copy a file.