@mattleibow 2020-04-09
I have been thinking about this PR... The delegates inline make is easier to maintain, but still a bit... well... not awesome.

I did like the way you did the static methods of `GetObject` in each type in https://github.com/mono/SkiaSharp/pull/1193. Maybe we could make the new, local `GetObject` just call the `SKObject` version:

```csharp
internal static GRContext GetObject (IntPtr ptr, bool owns = true, bool unrefExisting = true) =>
    SKObject.GetObject<GRContext>(handle, (h, o) => new GRContext(h, o), owns, unrefExisting);
```

This will also reduce the amount of IL generated for the thousands of delegates. I suppose that is less duplication and we just have to worry about the method signature, which is easy.

I know you have been working on this for several days and it is a pain to keep going back. So, thanks for sticking with it. Did you managed to run any perfs? It will be interesting to see how much the improvements are.