@Gillibald 2020-04-09
I am fine with redoing this. Your suggestion reduces the code noise so I am fine with it.

I wonder if we should get rid of the delegate as well if we introduce a helper method on each type. 

Creating object before this change took 3% after this change it was 1%. Destruction of objects still takes about 4% of all processing.

In general SkiaSharp had less impact on the overall processing time. 

I can run some benchmarks if you want.