@mattleibow 2020-04-09
Interesting that destruction takes more time... Do you have any idea where this overhead is? I will run some more profiling and see... Has there been any variance with types of objects/operations?