@Gillibald 2020-04-09
I am calling SKFontManager.MatchCharacter quite often. That produces `SKFontStyle` on every call. It is really bad to cache these one time used instances.