@mattleibow 2020-04-09
> That produces SKFontStyle on every call.

I saw that and now I cache that now: https://github.com/mono/SkiaSharp/blob/master/binding/Binding/SKFontStyle.cs#L55

But, did you try the overloads that take a parameter?
https://github.com/mono/SkiaSharp/blob/fc9f030a08eb158fb934774cdf54f8df939088c7/binding/Binding/SKFontManager.cs#L170