@Gillibald 2020-04-09
Y I will cache them on my own in the future still I don't see why SKFontStyle isn't a struct.