@Gillibald 2020-04-09
I will fix the unit tests tomorrow. When we agree on this solution I will fix coding style issues.