@mattleibow 2020-04-10
I did some benchmarks and things are looking like they are about 1.3x faster!

```
BenchmarkDotNet=v0.12.0, OS=Windows 10.0.18363
Intel Core i9-9980HK CPU 2.40GHz, 1 CPU, 16 logical and 8 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4150.0), X64 RyuJIT
  Job-VTGAFA : .NET Framework 4.8 (4.8.4150.0), X64 RyuJIT
  Job-MNUBEA : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT
  Job-XIDDXQ : Mono 6.8.0 (Visual Studio), X64
```

**BEFORE**

```
|  Method |       Runtime |     Mean |     Error |    StdDev |
|-------- |-------------- |---------:|----------:|----------:|
| TheTest |    .NET 4.7.2 | 1.369 us | 0.0211 us | 0.0197 us |
| TheTest | .NET Core 3.1 | 1.178 us | 0.0233 us | 0.0229 us |
| TheTest |          Mono | 2.811 us | 0.0536 us | 0.0550 us |
```
**AFTER**

```
|  Method |       Runtime |     Mean |     Error |    StdDev |
|-------- |-------------- |---------:|----------:|----------:|
| TheTest |    .NET 4.7.2 | 1.051 us | 0.0172 us | 0.0143 us |
| TheTest | .NET Core 3.1 | 0.940 us | 0.0181 us | 0.0169 us |
| TheTest |          Mono | 1.912 us | 0.0186 us | 0.0156 us |
```

**CODE**

```csharp
[SimpleJob(RuntimeMoniker.Mono)]
[SimpleJob(RuntimeMoniker.Net472)]
[SimpleJob(RuntimeMoniker.NetCoreApp31)]
public class Benchmark {
    [Benchmark]
    public object TheTest() {
        // the actual test
        var data = SKData.Create(1);

        data.Dispose();
        return data;
    }
}

public class Program {
    public static void Main(string[] args) {
        var summary = BenchmarkRunner.Run<Benchmark>();
    }
}
```