@Gillibald 2020-04-10
Hmm that isn't much but at least this will free us from reflection.