@mattleibow 2020-04-10
Not much, but it might add up on larger drawings. This is a 1.3x speed up on almost every static construction and many members. 
And this is just .NET desktop. Who knows what actually happens on iOS or other AOT systems.