@mattleibow 2020-04-11
Thanks for this PR, took a few runs but we got it all done and we now have a slightly faster SkiaSharp.