@Gillibald 2020-04-11
I am fine with redoing things to get the best results. Will start experimenting with the V2 bits soon so we get that stable soon.